+++ CHANGELOG.md
@@ -10,6 +10,18 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.13.0] - 2024-09-30
+
+### Added
+
+- Added `Marshaler.UnmarshalTypeFromMediaType()`.
+
+### Removed
+
+- Removed `Envelope.PortableName`. The `MediaType` field is now guaranteed to
+  include the portable name as a parameter.
+- Removed `Packet.PortableName()`.
+
 ## [0.12.2] - 2024-09-30
 
 ### Fixed